posts and comments cant be removed or edited if its not the user that made them

+++ frontend/src/components/PostIndexItem/index.js
@@ -53,7 +53,9 @@ const PostIndex = () => {
 
   const handleDeletePost = (e, postId) => {
     e.preventDefault();
-    return dispatch(deletePost(postId));
+    if (window.confirm("are you sure?")){
+        dispatch(deletePost(postId));
+    }
   }
 
   const handleEditPost = (e, post) => {
